requirements_dev.txt
# requirements_dev.txt includes the dependencies for running scripts on the CI
# e.g. documentation building and quality assurance (pre-commit).

pre-commit>=2.16.0
pytest-cov>=3.0.0
pytest-xdist>=2.5.0
mkdocs>=1.1
mkdocs-material>=4.6.3
mkdocs-git-revision-date-localized-plugin >= 1.2
mkdocs-glightbox >= 0.3.4
pymdown-extensions>=6.3
bokeh>=3.4.0,<4
mkdocstrings==0.18.0
Jinja2>=3.0
mkdocs-jupyter>=0.24
